restored one earlier YC job-ad-skipping technique
this code is kind of like a spam blocker in that it uses a bunch
of little hacks to prevent a particular message from getting
through. restored a hack based on things which link back to HN,
because a new job ad got past the other filters.

this has the side effect of blocking "ask HN" links, but that's
a pretty painless tradeoff for me personally